requirements.txt
shippo
django-naomi
django
djangorestframework
django-rest-auth
django-allauth
celery[redis]
django-model-utils
django-webpack-loader
django-js-reverse
django-import-export
django-ses
python-decouple
psycopg2
brotlipy
django-log-request-id
dj-database-url
gunicorn
whitenoise
psutil
ipython
sentry-sdk
requests
django_tables2
stripe
fcm_django
